+++ passkey-authenticator/src/authenticator/get_info.rs
@@ -1,4 +1,7 @@
-use passkey_types::ctap2::get_info::{Options, Response, Version};
+use passkey_types::{
+    ctap2::get_info::{Options, Response, Version},
+    webauthn::PublicKeyCredentialParameters,
+};
 
 use crate::{
     Authenticator, CredentialStore, UserValidationMethod, credential_store::DiscoverabilitySupport,
@@ -7,9 +10,9 @@ use crate::{
 impl<S: CredentialStore, U: UserValidationMethod> Authenticator<S, U> {
     /// Using this method, the host can request that the authenticator report a list of all
     /// supported protocol versions, supported extensions, AAGUID of the device, and its capabilities.
-    pub async fn get_info(&self) -> Response {
-        Response {
-            versions: vec![Version::FIDO_2_0, Version::U2F_V2],
+    pub async fn get_info(&self) -> Box<Response> {
+        Box::new(Response {
+            versions: vec![Version::FIDO_2_0, Version::FIDO_2_1],
             extensions: self.extensions.list_extensions(),
             aaguid: *self.aaguid(),
             options: Some(Options {
@@ -22,6 +25,14 @@ impl<S: CredentialStore, U: UserValidationMethod> Authenticator<S, U> {
             max_msg_size: None,
             pin_protocols: None,
             transports: Some(self.transports.clone()),
-        }
+            algorithms: Some(
+                self.algs
+                    .iter()
+                    .copied()
+                    .map(PublicKeyCredentialParameters::from)
+                    .collect(),
+            ),
+            ..Default::default()
+        })
     }
 }